split camera mixins into separate files
